+++ src/org/netbeans/modules/php/blade/editor/lexer/BladeTokenId.java
@@ -82,6 +82,9 @@ public enum BladeTokenId implements TokenId {
                 case PHP_BLADE_EXPRESSION:
                 case PHP_BLADE_ECHO_EXPR:
                 case PHP_BLADE_INLINE_CODE:
+                    if (1==1){
+                        //return null;
+                    }
                     Language<? extends TokenId> phpLanguage = PHPTokenId.languageInPHP();
                     return phpLanguage != null ? LanguageEmbedding.create(phpLanguage, 0, 0, false) : null;
 //                    return null;